Allow downloading multiple torrents to blackhole

+++ core/src/main/java/org/nzbhydra/downloading/TorrentHandlingWeb.java
@@ -1,25 +1,32 @@
 package org.nzbhydra.downloading;
 
 import com.google.common.io.Files;
-import org.nzbhydra.GenericResponse;
+import lombok.AllArgsConstructor;
+import lombok.Data;
 import org.nzbhydra.api.WrongApiKeyException;
 import org.nzbhydra.config.BaseConfig;
 import org.nzbhydra.config.ConfigProvider;
 import org.nzbhydra.config.NzbAccessType;
-import org.nzbhydra.misc.UserAgentMapper;
 import org.nzbhydra.searching.searchrequests.SearchRequest.SearchSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.io.File;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashSet;
 import java.util.Objects;
+import java.util.Set;
 
 @RestController
 public class TorrentHandlingWeb {
@@ -30,8 +37,6 @@ public class TorrentHandlingWeb {
     private NzbHandler nzbHandler;
     @Autowired
     private ConfigProvider configProvider;
-    @Autowired
-    private UserAgentMapper userAgentMapper;
 
     /**
      * Provides an internal access to torrents via GUID
@@ -53,7 +58,8 @@ public class TorrentHandlingWeb {
     @RequestMapping(value = "/gettorrent/user/{guid}", produces = "application/x-bittorrent")
     @Secured({"ROLE_USER"})
     public ResponseEntity<Object> downloadTorrentForUsers(@PathVariable("guid") long guid) throws InvalidSearchResultIdException{
-        return nzbHandler.getNzbByGuid(guid, configProvider.getBaseConfig().getSearching().getNzbAccessType(), SearchSource.INTERNAL).getAsResponseEntity();
+        NzbDownloadResult downloadResult = nzbHandler.getNzbByGuid(guid, configProvider.getBaseConfig().getSearching().getNzbAccessType(), SearchSource.INTERNAL);
+        return downloadResult.getAsResponseEntity();
     }
 
     /**
@@ -61,30 +67,44 @@ public class TorrentHandlingWeb {
      *
      * @return A {@link ResponseEntity} with the torrent content, a redirect to the actual indexer link or an error
      */
-    @RequestMapping(value = "/internalapi/saveTorrent/{guid}")
+    @RequestMapping(value = "/internalapi/saveTorrent", method= RequestMethod.PUT, produces = MediaType.APPLICATION_JSON_VALUE)
     @Secured({"ROLE_USER"})
-    public GenericResponse sentTorrentToBlackhole(@PathVariable("guid") long guid) throws InvalidSearchResultIdException{
-        NzbDownloadResult downloadResult = nzbHandler.getNzbByGuid(guid, NzbAccessType.PROXY, SearchSource.INTERNAL);
-        if (!downloadResult.isSuccessful()) {
-            return GenericResponse.notOk(downloadResult.getError());
-        }
+    public AddTorrentsResponse sentTorrentToBlackhole(@RequestBody Set<Long> searchResultIds) throws InvalidSearchResultIdException{
         if (!configProvider.getBaseConfig().getDownloading().getSaveTorrentsTo().isPresent()) {
             logger.error("Torrent black hole folder not set");
-            return GenericResponse.notOk("Torrent black hole folder not set");
+            return new AddTorrentsResponse(false, "Torrent black hole folder not set", Collections.emptyList(), searchResultIds);
         }
-        String sanitizedTitle = downloadResult.getTitle().replaceAll("[\\\\/:*?\"<>|]", "_");
-        if(!Objects.equals(sanitizedTitle, downloadResult.getTitle())) {
-            logger.info("Sanitized torrent title from '{}' to '{}'", downloadResult.getTitle(), sanitizedTitle);
+        Set<Long> addedIds = new HashSet<>();
+        for (Long guid : searchResultIds) {
+            NzbDownloadResult downloadResult = nzbHandler.getNzbByGuid(guid, NzbAccessType.PROXY, SearchSource.INTERNAL);
+            if (!downloadResult.isSuccessful()) {
+                return handleError("An error occurred while downloading torrent: " + downloadResult.getError(), searchResultIds, addedIds);
+            }
+            String sanitizedTitle = downloadResult.getTitle().replaceAll("[\\\\/:*?\"<>|]", "_");
+            if(!Objects.equals(sanitizedTitle, downloadResult.getTitle())) {
+                logger.info("Sanitized torrent title from '{}' to '{}'", downloadResult.getTitle(), sanitizedTitle);
+            }
+            File torrent = new File(configProvider.getBaseConfig().getDownloading().getSaveTorrentsTo().get(), sanitizedTitle + ".torrent");
+            if (torrent.exists()) {
+                logger.info("File {} already exists and will be skipped", torrent.getAbsolutePath());
+                continue;
+            }
+            try {
+                Files.write(downloadResult.getNzbContent().getBytes(), torrent);
+                logger.info("Saved torrent file to {}", torrent.getAbsolutePath());
+                addedIds.add(guid);
+            } catch (Exception e) {
+                logger.error("Error saving torrent file", e);
+                return handleError("Error saving torrent file: " + e.getMessage(), searchResultIds, addedIds);
+            }
         }
-        File torrent = new File(configProvider.getBaseConfig().getDownloading().getSaveTorrentsTo().get(), sanitizedTitle + ".torrent");
-        try {
-            Files.write(downloadResult.getNzbContent().getBytes(), torrent);
-            logger.info("Saved torrent file to {}", torrent.getAbsolutePath());
-        } catch (Exception e) {
-            logger.error("Error saving torrent file", e);
-            return GenericResponse.notOk("Error saving torrent file: " + e.getMessage());
-        }
-        return GenericResponse.ok();
+
+        return new AddTorrentsResponse(true, null, addedIds, Collections.emptySet());
+    }
+
+    protected AddTorrentsResponse handleError(String message, Set<Long> searchResultIds, Set<Long> addedIds) {
+        searchResultIds.removeAll(addedIds);
+        return new AddTorrentsResponse(false, message, addedIds, searchResultIds);
     }
 
     /**
@@ -100,7 +120,16 @@ public class TorrentHandlingWeb {
             throw new WrongApiKeyException("Wrong api key");
         }
 
-        return nzbHandler.getNzbByGuid(guid, baseConfig.getSearching().getNzbAccessType(), SearchSource.INTERNAL).getAsResponseEntity();
+        return nzbHandler.getNzbByGuid(guid, baseConfig.getSearching().getNzbAccessType(), SearchSource.API).getAsResponseEntity();
+    }
+
+    @Data
+    @AllArgsConstructor
+    public class AddTorrentsResponse {
+        private boolean successful;
+        private String message;
+        private Collection<Long> addedIds;
+        private Collection<Long> missedIds;
     }
 
 }
